requirements.txt
# Data Sources
nsepython>=2.97
# jugaad-data removed due to dependency conflicts (uses old click/beautifulsoup4)

# Greeks Calculation
scipy>=1.9.0
numpy>=1.23.0

# Broker
kiteconnect>=5.0.0

# Scheduling
APScheduler>=3.10.0
pytz>=2023.3

# Database & Config
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
requests>=2.28.0
loguru>=0.7.0

# Testing
pytest>=7.0.0
